I#47: Add focus input mode and receive new keystrokes.
Previously, VimFx dropped any keystrokes (besides Esc) that occured
when the user had focus in an editable area; this commit makes an
exception to this rule when the add-on is in the newly added focus
input mode and a tab character is read. Currently, the received
keystroke is logged; the full selection functionality will be available
in a later commit.

Additionally, slight refactoring is done around this event handling in
order to improve readability.

The 'gi' keystroke is acknowledged by the new focus input key handler,
which notifies the user which key the handler received via the console.