I#47: Add xpath and dom manipulation utils.
These utils include generic functions to create and evaluate xpaths in
addition to specific <input> tag xpath functionality. A method to
identify visible client rects of a parent element was also added.

These functions are exported from utils file but are not yet used
externally - this will happen in the upcoming commit. As such, there is
no user-facing functionality added in this commit.